+++ apps/landing/src/lib/pipes/get-pipe-docs-uri.ts
@@ -1,5 +1,5 @@
 import { getBasePipeName, getPipeVersion, PipeId } from "@pipe0/base";
 
 export function getPipeDocsURI(pipeId: PipeId) {
-  return `/docs/pipes/pipes-catalog/${getBasePipeName(pipeId)}/${getPipeVersion(pipeId as PipeId)}`;
+  return `/docs/pipe-catalog/${getBasePipeName(pipeId)}/${getPipeVersion(pipeId as PipeId)}`;
 }