+++ apps/landing/CHANGELOG.md
@@ -1,5 +1,19 @@
 # pipe0-landing
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @pipe0/doc-links@0.0.4
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @pipe0/doc-links@0.0.3
+
 ## 0.1.3
 
 ### Patch Changes